查看权限id、权限列表、移除权限、添加权限、分配权限、分配菜单

+++ src/main/java/top/integer/blog/enums/Permission.java
@@ -26,7 +26,7 @@ public enum Permission {
     /**
      * id
      */
-    public final int id;
+    public final long id;
     /**
      * 权限名称
      */
@@ -36,7 +36,7 @@ public enum Permission {
      */
     public final String description;
 
-    Permission(int id, String name, String description) {
+    Permission(long id, String name, String description) {
         this.id = id;
         this.name = name;
         this.description = description;
